Struct glean_build::Builder

pub struct Builder { /* private fields */ }

A Glean Rust bindings generator.

Implementations

impl Builder

pub fn with_output<S: Into<String>>(out_dir: S) -> Self

A Glean Rust bindings generator with the given output directory.

Use file and files to specify the input files.

pub fn file<S: Into<String>>(&mut self, file: S) -> &mut Self

Add a definition file, e.g. metrics.yaml or pings.yaml.

Can be called multiple times to add more files.

pub fn files<P>(&mut self, files: P) -> &mut Self
where P: IntoIterator, P::Item: Into<String>,

Add multiple definition files, e.g. metrics.yaml or pings.yaml.

pub fn generate(&self) -> Result<()>

Generate the Rust bindings.

The consumer must include the generated glean_metrics.rs file, e.g.:

include!(concat!(env!("OUT_DIR"), "/glean_metrics.rs"));